Title: 사칙연산 함수, Time: 49ms, Memory: 16MB, Status: Passed - Codetree

+++ 260202/사칙연산 함수/quadratic-operations-function.py
@@ -9,7 +9,7 @@ def cal(a, o, c):
     if o == '-':
         return a - c
     if o == '/':
-        return a / c
+        return int(a / c)
     if o == '*':
         return a * c
     return False
